Adjust encryptables controller
