+++ app/controllers/api/encryptables/file_controller.rb
@@ -4,6 +4,7 @@ class Api::Encryptables::FileController < Api::EncryptablesController
   self.permitted_attrs = [:filename, :description, :encryptable_credentials_id, :file]
 
   def create
+    require 'pry'; binding.pry unless $pstop
     build_encryptable_file
     authorize entry
     if entry.save
@@ -15,16 +16,6 @@ class Api::Encryptables::FileController < Api::EncryptablesController
   end
 
   private
-
-  def build_encryptable_file
-    filename = params[:file].original_filename
-    file = new_file(encryptable_credential, params[:description], filename)
-    file.cleartext_file = params[:file].read
-
-    file.encrypt(plaintext_team_password(team))
-
-    instance_variable_set(:"@#{ivar_name}", file)
-  end
 
   def new_file(parent_encryptable, description, name)
     Encryptable::File.new(folder_id: parent_encryptable.folder_id,
