+++ frontend/app/templates/components/encryptable/show.hbs
@@ -1,8 +1,8 @@
 {{#if this.isEncryptableEditing}}
   <Encryptable::Form @encryptable={{@encryptable}} @title={{t "encryptables.edit.title"}} @onAbort={{this.toggleEncryptableEdit}}/>
 {{/if}}
-{{#if this.isFileEntryCreating}}
-  <EncryptableFile::Form @class="modal_file_entry" @encryptableCredential={{@encryptable}} @title={{t "file_entries.new.title"}} @onAbort={{this.toggleFileEntryNew}}/>
+{{#if this.isFileCreating}}
+  <EncryptableFile::Form @class="modal_file_entry" @encryptableCredential={{@encryptable}} @title={{t "file_entries.new.title"}} @onAbort={{this.toggleFileNew}}/>
 {{/if}}
 <div class="container px-5 pt-4 h-100 bg-white pl-none encryptable-container-smartphone mb-15">
   <div class="row mb-3">
